feat: implement onboarding completion functionality and refactor related services and routes
