+++ src/routers/user.route.js
@@ -1,6 +1,6 @@
 import {Router} from "express"
 
-import {  createUserController, deleteUserByIdController, getAllUsersController, getUserByIdController, getUserByUserNameController, updateUserByUserNameController } from "../controllers/user.controller.js"
+import {  createUserController, deleteUserByIdController, getAllUsersController, getUserByIdController, getUserByUserNameController, updateUserByUserNameController,markOnboardingCompleteController } from "../controllers/user.controller.js"
 import { userByIdValidation, userByUserNameValidation, userValidation } from "../middleware/validation.js"
 
 
@@ -13,6 +13,7 @@ userRouter.get("/users/:userName",userByUserNameValidation ,getUserByUserNameCon
 userRouter.get("/users/:id",userByIdValidation,getUserByIdController)
 userRouter.put("/users/:userName",userByUserNameValidation ,updateUserByUserNameController)
 userRouter.delete("/users/:id", userByIdValidation,deleteUserByIdController)
+userRouter.patch("/users/:id/onboarding/complete", userByIdValidation, markOnboardingCompleteController)
